refactor: model 패키지명을 domain으로 수정
