+++ src/main/java/blackjack/domain/Cards.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.domain;
 
 import java.util.Collections;
 import java.util.List;
