+++ src/main/java/blackjack/domain/Deck.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.domain;
 
 import java.util.ArrayList;
 import java.util.Arrays;
