+++ src/main/java/blackjack/domain/HitCommand.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.domain;
 
 import blackjack.exception.ErrorMessage;
