+++ src/main/java/blackjack/domain/Name.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.domain;
 
 import blackjack.exception.ErrorMessage;
