+++ src/main/java/blackjack/domain/Score.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.domain;
 
 import java.util.List;
