+++ src/main/java/blackjack/view/OutputView.java
@@ -1,10 +1,10 @@
 package blackjack.view;
 
-import blackjack.model.*;
-import blackjack.model.Dealer;
-import blackjack.model.Participant;
-import blackjack.model.Player;
-import blackjack.model.Players;
+import blackjack.domain.*;
+import blackjack.domain.Dealer;
+import blackjack.domain.Participant;
+import blackjack.domain.Player;
+import blackjack.domain.Players;
 
 import java.util.LinkedHashMap;
 import java.util.List;
