+++ src/test/java/blackjack/domain/DeckTest.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
